refactor(text-to-image): moves "Config/Dynamic/exports.ts" to "Config/Dynamic/fetch.ts"
