+++ src/features/TextToImage/Config/Dynamic/index.ts
@@ -1,1 +1,1 @@
-export * from './exports.ts';
+export * from './fetch.ts';
